requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
openai==1.12.0
python-docx==1.1.0
PyPDF2==3.0.1
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-mock==3.12.0
